requirements.txt
Django==4.2.1
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.2
django-debug-toolbar==4.0.0
django-cors-headers==4.0.0